Protocol improvements, added a Makefile, other misc changes.

+++ moist.h
@@ -11,6 +11,6 @@
 #define MOIST_COLLECTOR_PIN	3
 #define MOIST_MAX_VALUE		(const uint8_t)((1l<<(sizeof(uint8_t)*8l))-1)
 
-extern uint8_t moist_calc();
+extern uint16_t moist_calc();
 
 #endif
